Refactor resource cache: Introduce HashList and HashItem types, and remove resourceupdate.go
